code structured gonna add an express route and setup functions in that
