+++ src/components/Custom.js
@@ -1,0 +1,4 @@
+const Custom = () => {
+    return <p className="text-red-500">customization available</p>;
+  };
+  export default Custom
