+++ src/components/NotAvailable.js
@@ -1,0 +1,4 @@
+const NotAvailable = () => {
+    return <div className="text-red-500">Not available</div>;
+  };
+  export default NotAvailable
